// src/gen/custom/store_pb.ts
// @generated by protoc-gen-es v2.12.0 with parameter "target=ts"
// @generated from file custom/store.proto (syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv2";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv2";
import type { Entity } from "../entity_pb";
import { file_entity } from "../entity_pb";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file custom/store.proto.
 */
export const file_custom_store: GenFile = /*@__PURE__*/
  fileDesc("ChJjdXN0b20vc3RvcmUucHJvdG8iKQoLRGF0YVN0b3JhZ2USGgoCdjEYASABKAsyDi5EYXRhU3RvcmFnZVYxIjwKDURhdGFTdG9yYWdlVjESGQoIZW50aXRpZXMYASADKAsyBy5FbnRpdHkSEAoIc3luY191cmwYAiABKAliBnByb3RvMw", [file_entity]);

/**
 * @generated from message DataStorage
 */
export type DataStorage = Message<"DataStorage"> & {
  /**
   * @generated from field: DataStorageV1 v1 = 1;
   */
  v1?: DataStorageV1 | undefined;
};

/**
 * Describes the message DataStorage.
 * Use `create(DataStorageSchema)` to create a new message.
 */
export const DataStorageSchema: GenMessage<DataStorage> = /*@__PURE__*/
  messageDesc(file_custom_store, 0);

/**
 * @generated from message DataStorageV1
 */
export type DataStorageV1 = Message<"DataStorageV1"> & {
  /**
   * @generated from field: repeated Entity entities = 1;
   */
  entities: Entity[];

  /**
   * @generated from field: string sync_url = 2;
   */
  syncUrl: string;
};

/**
 * Describes the message DataStorageV1.
 * Use `create(DataStorageV1Schema)` to create a new message.
 */
export const DataStorageV1Schema: GenMessage<DataStorageV1> = /*@__PURE__*/
  messageDesc(file_custom_store, 1);